# backend springboot/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Attendance QR API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Err401"
...
  /auth/register:
    post:
      tags: [ Auth ]
      security: [ ]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/Err400"
        "409":
          $ref: "#/components/responses/Err409"
  /groups:
    post:
      tags: [Groups]
      summary: Create group (caller becomes OWNER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
        "409":
          $ref: "#/components/responses/Err409"
        "400":
          $ref: "#/components/responses/Err400"

    get:
      tags: [Groups]
      summary: List my groups
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageGroupResponse"
...
  /groups/join:
    post:
      tags: [ Members ]
      summary: Join group by joinCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        "400":
          $ref: "#/components/responses/Err400"
        "401":
          $ref: "#/components/responses/Err401"
        "404":
          $ref: "#/components/responses/Err404"
        "409":
          $ref: "#/components/responses/Err409"
  /groups/{groupId}/members:
    get:
      tags: [ Members ]
      summary: List members (OWNER/CO_HOST)
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageMemberResponse"
        "401":
          $ref: "#/components/responses/Err401"
        "403":
          $ref: "#/components/responses/Err403"
        "404":
          $ref: "#/components/responses/Err404"
  /groups/{groupId}/members/{userId}:
    patch:
      tags: [ Members ]
      summary: Member action (approve/reject/remove/promote/demote/transfer ownership)
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberActionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberResponse"
        "400":
          $ref: "#/components/responses/Err400"
        "401":
          $ref: "#/components/responses/Err401"
        "403":
          $ref: "#/components/responses/Err403"
        "404":
          $ref: "#/components/responses/Err404"
        "409":
          $ref: "#/components/responses/Err409"
        "422":
          $ref: "#/components/responses/Err422"
  /groups/{groupId}/members/me:
    delete:
      tags: [ Members ]
      summary: Leave group (MEMBER/CO_HOST)
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Err401"
        "404":
          $ref: "#/components/responses/Err404"
        "409":
          $ref: "#/components/responses/Err409"
        "422":
          $ref: "#/components/responses/Err422"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Sort:
      name: sort
      in: query
      schema:
        type: string
        example: "createdAt,desc"
  responses:
    Err400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            validation:
              value:
                status: 400
                code: VALIDATION_ERROR
                message: Invalid request
    Err401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            unauthorized:
              value:
                status: 401
                code: UNAUTHORIZED
                message: Missing or invalid token
    Err403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            forbidden:
              value:
                status: 403
                code: FORBIDDEN
                message: Not allowed
    Err404:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            notFound:
              value:
                status: 404
                code: NOT_FOUND
                message: Resource not found
    Err409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            conflict:
              value:
                status: 409
                code: CONFLICT
                message: Conflict
    Err422:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          examples:
            rule:
              value:
                status: 422
                code: BUSINESS_RULE_VIOLATION
                message: Rule violation
  schemas:
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        path:
          type: string
        traceId:
          type: string
      required: [timestamp, status, code, message, path]
    PlatformRole:
      type: string
      enum: [ADMIN, USER]
...
    ApprovalMode:
      type: string
      enum: [AUTO, MANUAL]
    GroupStatus:
      type: string
      enum: [ACTIVE, ARCHIVED]
    MemberRole:
      type: string
      enum: [OWNER, CO_HOST, MEMBER]
    MemberStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, REMOVED]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 190
        password:
          type: string
          minLength: 8
          maxLength: 200
      required: [email, password]
    LoginUserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        platformRole:
          $ref: "#/components/schemas/PlatformRole"
      required: [id, email, fullName, platformRole]
    TokenResponse:
      type: object
      properties:
        tokenType:
          type: string
          example: Bearer
        accessToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/LoginUserSummary"
      required: [ tokenType, accessToken, accessTokenExpiresAt, user ]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 190
        password:
          type: string
          minLength: 8
          maxLength: 200
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        userCode:
          type: string
          maxLength: 40
        deviceId:
          type: string
          maxLength: 120
      required: [ email, password, fullName ]
    RegisterResponse:
        type: object
        properties:
          tokenType:
            type: string
            example: Bearer
          accessToken:
            type: string
          accessTokenExpiresAt:
            type: string
            format: date-time
          user:
            $ref: "#/components/schemas/LoginUserSummary"
          firstLogin:
            type: boolean
        required: [ tokenType, accessToken, accessTokenExpiresAt, user, firstLogin ]
...
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 150
        code:
          type: string
          maxLength: 20
        joinCode:
          type: string
          minLength: 6
          maxLength: 16
        description:
          type: string
          maxLength: 1000
        semester:
          type: string
          maxLength: 30
        room:
          type: string
          maxLength: 80
        approvalMode:
          $ref: "#/components/schemas/ApprovalMode"
        allowAutoJoinOnCheckin:
          type: boolean
          default: false
      required: [name, code, approvalMode]
...
    GroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        code:
          type: string
          maxLength: 20
        joinCode:
          type: string
          maxLength: 16
        description:
          type: string
          maxLength: 1000
        semester:
          type: string
          maxLength: 30
        room:
          type: string
          maxLength: 80
        approvalMode:
          $ref: "#/components/schemas/ApprovalMode"
        allowAutoJoinOnCheckin:
          type: boolean
        status:
          $ref: "#/components/schemas/GroupStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, ownerUserId, name, code, joinCode, approvalMode, status, createdAt, updatedAt]
    JoinGroupRequest:
      type: object
      properties:
        joinCode:
          type: string
          minLength: 6
          maxLength: 16
      required: [joinCode]
    MemberActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [ APPROVE, REJECT, REMOVE, PROMOTE, DEMOTE, TRANSFER_OWNERSHIP ]
        note:
          type: string
          maxLength: 500
      required: [ action ]
    MemberResponse:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/MemberRole"
        memberStatus:
          $ref: "#/components/schemas/MemberStatus"
        joinedAt:
          type: string
          format: date-time
          nullable: true
        invitedBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        removedAt:
          type: string
          format: date-time
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        fullName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required: [ groupId, userId, role, memberStatus, createdAt, updatedAt ]
...
    PageGroupResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GroupResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
      required: [items, page, size, totalElements, totalPages]
    PageMemberResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MemberResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
      required: [items, page, size, totalElements, totalPages]
...